backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25
django-cors-headers>=4.6,<5
mssql-django>=1.5,<2
pyodbc>=5.1,<6
celery>=5.4,<6
django-celery-beat>=2.7,<3
redis>=5.2,<6
django-redis>=5.4,<6
